reorganized files around HubContext
